+++ functions/index.js
@@ -28,6 +28,7 @@ exports.linkUserToPerson = require('./onUserCreate').linkUserToPerson
 exports.helloWorld = require('./helloWorld').helloWorld
 exports.logStartupNameChange = require('./logStartupNameChange').logStartupNameChange
 exports.onEventCreated = require('./onEventCreated').onEventCreated
+exports.joiningPremoney = require('./joiningPremoney').joiningPremoney
 
 // ✅ Export v2 callable functions
 exports.getHostEvents = onCall({ cors: true }, getHostEvents)